RAG/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv>=1.0.1
pydantic>=2.7.4,<3.0.0
python-multipart==0.0.6

# RAG dependencies
# Note: torch и torchvision устанавливаются отдельно в Dockerfile (CPU версия)
sentence-transformers==2.2.2
huggingface-hub==0.14.1
transformers==4.30.2
chromadb==0.4.18
langchain-text-splitters==0.0.1
python-docx==1.1.0
markitdown==0.1.4

# LLM
gigachat==0.1.12

# Utilities
numpy==1.24.3

